Update xml-manage.md
In the "Add or change itemprop global attribute for products" section, the XML and HTML code and examples markup was missing some "```". 
This caused the "To add the `itemprop` attribute..." text to appear in the same block as the XML and HTML examples. 
This PR corrects the markup to separate the explanation text and code examples.